feat: fix bug in fast path
Before this bug fix, pixelmatch.py would report that the two new test
images in this commit are identical.  This is not the case, only a bit
more than the upper quarter of the images are the same.

The images were rendered by Chrome version 80.0.3987.132 running on
Windows 7 and Windows 10 using the Segoe UI font.  On the Windows 7
machine the font had been updated to be the same as on Windows 10
by following steps similar to the ones found here:

https://www.quora.com/Is-it-possible-to-use-the-Windows-10-version-of-Segoe-UI-font-on-my-Windows-7-How

Unfortunately pixelmatch is only able to identify about 61% of the
pixels that differ between the two images as a result of antialiasing
(23903 pixels out of 38893), but in this case 100% of the pixels that
differ are due to antialiasing.